+++ examples/api-tests/src/preferences.spec.js
@@ -66,16 +66,13 @@ describe('Preferences', function () {
     }
 
     async function deleteAllValues() {
-        return setValueTo('', [], undefined);
+        return setValueTo(undefined);
     }
 
     /**
-     * 
-     * @param {string} key 
-     * @param {string[]} path 
-     * @param {any} value 
+     * @param {any} value - A JSON value to write to the workspace preference file.
      */
-    async function setValueTo(key, path, value) {
+    async function setValueTo(value) {
         const reference = await modelService.createModelReference(uri);
         if (reference.object.dirty) {
             await reference.object.revert();
@@ -83,7 +80,7 @@ describe('Preferences', function () {
         /** @type {import ('@theia/preferences/lib/browser/folder-preference-provider').FolderPreferenceProvider} */
         const provider = Array.from(folderPreferences['providers'].values()).find(candidate => candidate.getConfigUri().isEqual(uri));
         assert.isDefined(provider);
-        await provider['doSetPreference'](key, path, value);
+        await provider['doSetPreference']('', [], value);
         reference.dispose();
     }
 
@@ -105,7 +102,7 @@ describe('Preferences', function () {
             let content = '';
             try { content = JSON.parse(contentBeforehand); } catch { }
             // Use the preference service because its promise is guaranteed to resolve after the file change is complete.
-            await setValueTo('', [], content);
+            await setValueTo(content);
         }
     });