+++ packages/editor/src/common/language-selector.ts
@@ -13,6 +13,7 @@
 //
 // SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
 // *****************************************************************************
+
 import { match as matchGlobPattern } from '@theia/core/lib/common/glob';
 
 export interface RelativePattern {